requirements.txt
fasthtml>=0.0.1
google-generativeai>=0.8.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
